moosync_edk::api
Module extension_api

Functions

add_playlist

Adds a new playlist to the main app.

add_songs

Adds a list of songs to the main app.

add_to_playlist

Adds a song to a playlist.

gen_hash

get_current_song

Retrieves the current song being played.

get_entity

get_player_state

Retrieves the current state of the player.

get_preference

Retrieves a preference value based on the provided data.

get_queue

Retrieves the current playback queue.

get_secure

Retrieves a secure preference value based on the provided data.

get_song

Retrieves a list of songs based on the provided options.

get_system_time

get_time

Retrieves the current playback time.

get_volume

Retrieves the current volume level.

open_external_url

Opens an external URL.

open_sock

read_sock

register_oauth

Registers an OAuth token with the main app.

remove_song

Removes a song from the main app.

set_preference

Sets a preference value based on the provided data.

set_secure

Sets a secure preference value based on the provided data.

update_accounts

Updates the list of accounts in the main app.

update_song

Updates a song in the main app.

write_sock