moosync_edk::api

Trait Accounts

pub trait Accounts {
    // Provided methods
    fn get_accounts(&self) -> MoosyncResult<Vec<ExtensionAccountDetail>> { ... }
    fn perform_account_login(
        &self,
        args: AccountLoginArgs,
    ) -> MoosyncResult<String> { ... }
    fn oauth_callback(&self, code: String) -> MoosyncResult<()> { ... }
}

Trait for handling account-related events.

Provided Methods

fn get_accounts(&self) -> MoosyncResult<Vec<ExtensionAccountDetail>>

Called when the main app requests the list of accounts.

fn perform_account_login(&self, args: AccountLoginArgs) -> MoosyncResult<String>

Called when the main app requests to perform an account login.

fn oauth_callback(&self, code: String) -> MoosyncResult<()>

Called when the main app provides an OAuth callback code.