moosync_edk::api
Trait ContextMenu

pub trait ContextMenu {
    // Provided methods
    fn get_song_context_menu(
        &self,
        songs: Vec<Song>,
    ) -> MoosyncResult<Vec<ContextMenuReturnType>> { ... }
    fn get_playlist_context_menu(
        &self,
        playlist: QueryablePlaylist,
    ) -> MoosyncResult<Vec<ContextMenuReturnType>> { ... }
    fn on_context_menu_action(&self, action: String) -> MoosyncResult<()> { ... }
}

Trait for handling context menu-related events.

Provided Methods

fn get_song_context_menu( &self, songs: Vec<Song>, ) -> MoosyncResult<Vec<ContextMenuReturnType>>

Called when the main app requests the context menu for songs.

fn get_playlist_context_menu( &self, playlist: QueryablePlaylist, ) -> MoosyncResult<Vec<ContextMenuReturnType>>

Called when the main app requests the context menu for a playlist.

fn on_context_menu_action(&self, action: String) -> MoosyncResult<()>

Called when the main app performs an action from the context menu.