moosync_edk::api

Trait Provider

pub trait Provider {

    // Required method
    fn get_provider_scopes(&self) -> MoosyncResult<Vec<ExtensionProviderScope>>;

    // Provided methods
    fn get_playlists(&self) -> MoosyncResult<Vec<Playlist>> { ... }
    fn get_playlist_content(
        &self,
        id: String,
        next_page_token: Option<String>,
    ) -> MoosyncResult<SongsWithPageTokenReturnType> { ... }
    fn get_playlist_from_url(
        &self,
        url: String,
    ) -> MoosyncResult<Option<Playlist>> { ... }
    fn get_playback_details(
        &self,
        song: Song,
    ) -> MoosyncResult<PlaybackDetailsReturnType> { ... }
    fn search(&self, term: String) -> MoosyncResult<SearchResult> { ... }
    fn get_recommendations(&self) -> MoosyncResult<Vec<Song>> { ... }
    fn get_song_from_url(&self, url: String) -> MoosyncResult<Option<Song>> { ... }
    fn handle_custom_request(
        &self,
        url: String,
    ) -> MoosyncResult<CustomRequestReturnType> { ... }
    fn get_artist_songs(
        &self,
        artist: Artist,
        next_page_token: Option<String>,
    ) -> MoosyncResult<SongsWithPageTokenReturnType> { ... }
    fn get_album_songs(
        &self,
        album: Album,
        next_page_token: Option<String>,
    ) -> MoosyncResult<SongsWithPageTokenReturnType> { ... }
    fn get_song_from_id(&self, id: String) -> MoosyncResult<Option<Song>> { ... }
    fn scrobble(&self, song: Song) -> MoosyncResult<()> { ... }
    fn get_lyrics(&self, song: Song) -> MoosyncResult<String> { ... }

}

Trait for handling provider-related events.

Required Methods

fn get_provider_scopes(&self) -> MoosyncResult<Vec<ExtensionProviderScope>>

Called when the main app requests the provider scopes.

Provided Methods

fn get_playlists(&self) -> MoosyncResult<Vec<Playlist>>

Called when the main app requests the list of playlists.

fn get_playlist_content( &self, id: String, next_page_token: Option<String>, ) -> MoosyncResult<SongsWithPageTokenReturnType>

Called when the main app requests the content of a specific playlist.

fn get_playlist_from_url(&self, url: String) -> MoosyncResult<Option<Playlist>>

Called when the main app requests a playlist from a URL.

fn get_playback_details( &self, song: Song, ) -> MoosyncResult<PlaybackDetailsReturnType>

Called when the main app requests playback details for a song.

fn search(&self, term: String) -> MoosyncResult<SearchResult>

Called when the main app performs a search.

fn get_recommendations(&self) -> MoosyncResult<Vec<Song>>

Called when the main app requests recommendations.

fn get_song_from_url(&self, url: String) -> MoosyncResult<Option<Song>>

Called when the main app requests a song from a URL.

fn handle_custom_request( &self, url: String, ) -> MoosyncResult<CustomRequestReturnType>

Called when the main app handles a custom request.

fn get_artist_songs( &self, artist: Artist, next_page_token: Option<String>, ) -> MoosyncResult<SongsWithPageTokenReturnType>

Called when the main app requests songs of a specific artist.

fn get_album_songs( &self, album: Album, next_page_token: Option<String>, ) -> MoosyncResult<SongsWithPageTokenReturnType>

Called when the main app requests songs of a specific album.

fn get_song_from_id(&self, id: String) -> MoosyncResult<Option<Song>>

Called when the main app requests a song from an ID.

fn scrobble(&self, song: Song) -> MoosyncResult<()>

Called when the main app requests to scrobble a song.

fn get_lyrics(&self, song: Song) -> MoosyncResult<String>

Called when the main app requests lyrics for a song.